// admin-ui/api/generated/models/Import1688SourceRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Import1688SourceRequest = {
    url: string;
    scene?: (string | null);
    optimize_title?: (boolean | null);
    overwrite_title?: boolean;
    overwrite_dimensions?: boolean;
    selected_sku_id?: (string | null);
    selected_sku_price?: (number | null);
};