// admin-ui/api/generated/models/OzonSyncRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OzonSyncRequest = {
    date_from?: (string | null);
    date_to?: (string | null);
    client_id?: (string | null);
    api_key?: (string | null);
    limit?: number;
    max_pages?: number;
};